Moved git cmd to gutils package, and updated usage. Updated to yaml config for vx init
