+++ .gitignore
@@ -36,5 +36,3 @@ vx
 
 
 dist/
-# vx config - vexal.io 
-.vx
