+++ pkg/gutils/git_cmd.go
@@ -1,4 +1,4 @@
-package pr
+package gutils
 
 import (
 	"fmt"
